Added Swagger and Swagger ui integration

+++ service/src/main/java/com/iberianmotorsports/service/service/implementation/UserServiceImpl.java
@@ -1,5 +1,7 @@
 package com.iberianmotorsports.service.service.implementation;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.iberianmotorsports.service.ErrorMessages;
 import com.iberianmotorsports.service.model.User;
 import com.iberianmotorsports.service.repository.OpenIdRepository;
@@ -13,6 +15,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -32,13 +35,35 @@ public class UserServiceImpl implements UserService {
     @Autowired
     OpenIdRepository openIdRepository;
 
+
+    private RestTemplate restTemplate = new RestTemplate();
+
     static final Pageable pageable = PageRequest.of(0,10);
 
 
     @Override
-    public User saveUser(User user) {
-        user = findSteamInfo(user);
+    public User saveUser(Long steamId) {
+        if (isAlreadyInDatabase(steamId)) throw new ServiceException(ErrorMessages.DUPLICATE_USER.getDescription());
+        User user = getPlayerSummary(String.valueOf(steamId));
         return userRepository.save(user);
+    }
+
+
+    public User getPlayerSummary(String steamId) {
+        String apiKey = "6614BF6FC7820DF1DD2C875DB66C5D82";
+        String url = "http://api.steampowered.com/ISteamUser/GetPlayerSummaries/v0002/" + "?key=" + apiKey + "&steamids=" + steamId;
+
+        String response = restTemplate.getForObject(url, String.class);
+
+        ObjectMapper mapper = new ObjectMapper();
+        User user = null;
+        try {
+            user = mapper.readValue(response, User.class);
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException(e);
+        }
+
+        return user;
     }
 
     @Override
@@ -60,6 +85,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(Long steamId) {
+        if(!isAlreadyInDatabase(steamId)) {
+            throw new ServiceException(ErrorMessages.USER_NOT_IN_DB.getDescription());
+        }
         userRepository.deleteById(steamId);
     }
